✅ Feat: restructure app to focus on hero landing page
