+++ package.json
@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "@heroui/button": "^2.2.23-beta.2",
+    "@heroui/card": "^2.2.24",
     "@heroui/code": "^2.2.17-beta.5",
     "@heroui/dropdown": "^2.3.23-beta.2",
     "@heroui/input": "^2.4.23-beta.4",
