+++ src/App.tsx
@@ -5,10 +5,10 @@ import { CookieConsent } from "./components/cookie-consent";
 import { PageNotFound } from "./pages/404";
 
 import IndexPage from "@/pages/index";
-import DocsPage from "@/pages/docs";
+/* import DocsPage from "@/pages/docs";
 import PricingPage from "@/pages/pricing";
 import BlogPage from "@/pages/blog";
-import AboutPage from "@/pages/about";
+import AboutPage from "@/pages/about"; */
 import "./App.css";
 
 function App() {
@@ -17,10 +17,10 @@ function App() {
       <CookieConsent />
       <Routes>
         <Route element={<IndexPage />} path="/" />
-        <Route element={<DocsPage />} path="/docs" />
+        {/* <Route element={<DocsPage />} path="/docs" />
         <Route element={<PricingPage />} path="/pricing" />
         <Route element={<BlogPage />} path="/blog" />
-        <Route element={<AboutPage />} path="/about" />
+        <Route element={<AboutPage />} path="/about" /> */}
         <Route element={<PageNotFound />} path="*" />
       </Routes>
     </CookieConsentProvider>
